[GEOT-6769] Refact ClippingFeatureCollection
